requirements.txt
streamlit==1.28.1
asyncio
requests
beautifulsoup4
spacy
nltk
transformers
torch
sentence-transformers
textstat
matplotlib
plotly
pandas
numpy
pillow
opencv-python
pexels-api-py
python-dotenv
aiofiles
asyncpg
celery
redis
schedule
wordcloud
seaborn
scikit-learn